+++ src/approx/src/ApproxQuadraticSpline.cpp
@@ -283,7 +283,7 @@ HRESULT IfaceCalling ApproxQuadraticSpline::GetBounds(TGlucoseLevelBounds *bound
     if (maskedCount < 1)
         return S_FALSE;
 
-    uint32_t skipSides[2] = { 0, maskedCount };
+    size_t skipSides[2] = { 0, maskedCount };
 
     CFindMaskedBounds fb(values, appCurrentTestMask, skipSides);
     tbb::parallel_reduce(tbb::blocked_range<size_t>(0, maskedCount), fb);
